print return code in warning message after exec as 8-bit singed value
Do not print "exit code 255", but "exit code -1"

+++ nsd/exec.c
@@ -188,7 +188,7 @@ Ns_WaitForProcessStatus(pid_t pid, int *exitcodePtr, int *waitstatusPtr)
 
         if (exitcode != 0) {
             Ns_Log(Warning, "process %d exited with nonzero exit code: %d",
-                   pid, exitcode);
+                   pid, (int8_t)exitcode);
         }
         if (exitcodePtr != NULL) {
             *exitcodePtr = exitcode;
